Updated readme with latest CLI functionality.
First pass at cli test documentation shows that things are getting
complicated. It would be good to have broader sections, as well as
schematics for the overall structure of this workflow. For example:
what's the difference between an image and a container, how do we track
them, where is your data and the command saved, etc.