Added functioning version of run_dgp_ensemble.
Added a script that does the input managemet for run_dgp_ensemble.
Some notes: it's still very difficult to figure out when something has gone wrong with the current setup.
it would be a good use of time to take your new logging functions, and
replace the old ones in ncap utils with them.